some inset fixes + other fixes for warnings from cxx
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@1046 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/Bullet.h
@@ -106,11 +106,11 @@ private:
 	static string const bulletEntry(short int, short int);
 	
 	///
-	short font;
+	int font;
 	///
-	short character;
+	int character;
 	///
-	short size;
+	int size;
 	
 	// size, character and font are array indices to access 
 	// the predefined arrays of LaTeX equivalent strings.
